+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased][unreleased]
 
+## Added
+- add received timestamp to html email view
+
 <a id="1_36_3"></a>
 
 ## [1.36.3] - 2023-04-17